settings page mobile support

+++ Assets/AirshipPackages/@Easy/Core/Shared/MainMenu/Components/Settings/DeleteAccountButton.ts
@@ -1,13 +1,20 @@
+import { Dependency } from "@Easy/Core/Shared/Flamework";
 import { AirshipUrl } from "@Easy/Core/Shared/Util/AirshipUrl";
 import { Bin } from "@Easy/Core/Shared/Util/Bin";
 import { CanvasAPI } from "@Easy/Core/Shared/Util/CanvasAPI";
+import { MainMenuSingleton } from "../../Singletons/MainMenuSingleton";
 
 export default class DeleteAccountButton extends AirshipBehaviour {
 	private bin = new Bin();
 
 	override OnEnable(): void {
 		CanvasAPI.OnClickEvent(this.gameObject, () => {
-			task.spawn(() => {
+			task.spawn(async () => {
+				const confirmed = await Dependency<MainMenuSingleton>().ShowConfirmModal(
+					"Delete Account",
+					"Are you sure you want to delete your account? This cannot be undone.",
+				);
+				if (!confirmed) return;
 				const res = InternalHttpManager.DeleteAsync(AirshipUrl.GameCoordinator + "/users/self");
 				if (res.error) {
 					error(res.error);
